US CFTC futures position Bloomberg tickers

The dataset includes Bloomberg tickers for the US Commodity Futures Trading Commission (CFTC)'s 'legacy', 'disaggregated', 'supplemental' and 'traders in financial futures' reports. See 'Details' section below.

tickers_CFTC

Format

A data.table. Columns include:

name: name of the futures series.

asset class: asset class for the underlying futures series ('climate', 'commodity', 'equity', 'fixed income' or 'currency').

active contract ticker: active contract ticker for the underlying futures series.

MIC: ISO 10383 Codes for exchanges and market identification (MIC). Identifies the exchange where the corresponding futures series trades.

format: CFTC report format ('legacy', 'disaggregated', 'supplemental' or 'traders in financial futures').

underlying: underlying instrument ('futures only' or 'futures & options').

unit: counting unit (number of 'contracts', 'traders' or 'total').

participant: CFTC's trader classification. Report specific:

legacy: 'commercial', 'non-commercial', 'non-reportable', 'total'.

disaggregated: 'managed money', 'producer/merchant/processor/user', 'swap dealers', 'other reportables'.

supplemental: 'commercial - non-CIT', 'non-commercial - non-CIT', 'index traders - non-CIT', 'index traders'.

traders in financial futures: 'asset manager/institutional', 'dealer/intermediary', 'leveraged funds', 'other reportables'.

position: trader positions. Participant specific.

commercial: 'long', 'short', 'net'.

non-commercial: 'long', 'short', 'net', 'spreading'.

non-reportable: 'long', 'short', 'net'.

total: 'long', 'short', 'net', 'open interest', 'total'.

managed money: 'long', 'short', 'net', 'spreading'.

producer/merchant/processor/user: 'long', 'short', 'net'.

swap dealers: 'long', 'short', 'net', 'spreading'.

other reportables: 'long', 'short', 'net', 'spreading'.

commercial - non-CIT: 'long', 'short', 'net'.

non-commercial - non-CIT: 'long', 'short', 'net', 'spreading'.

index traders - non-CIT: 'long', 'short'.

index traders: 'long', 'short', 'net'.

asset manager/institutional: 'long', 'short', 'net', 'spreading'.

dealer/intermediary: 'long', 'short', 'net', 'spreading'.

leveraged funds: 'long', 'short', 'net', 'spreading'.

other reportables: 'long', 'short', 'net', 'spreading'.

ticker: corresponding Bloomberg ticker.

Source

http://www.cftc.gov/MarketReports/CommitmentsofTraders/index.htm

http://www.cftc.gov/MarketReports/CommitmentsofTraders/ExplanatoryNotes/index.htm

http://www.cftc.gov/idc/groups/public/@commitmentsoftraders/documents/file/tfmexplanatorynotes.pdf

Details

The Commitments of Traders (COT) reports provide a breakdown of each Tuesday's open interest for markets in which 20 or more traders hold positions equal to or above the reporting levels established by the CFTC. A trader must report his or her position if at the daily market close, their position is at or above the CFTC's reporting level in any futures month or option expiration. A trader is determined to be commercial or non-commercial using the following rational: all traders' reported futures positions in a commodity are classified as commercial if the trader uses futures contracts in that particular commodity for hedging as defined in the CFTC's regulations (1.3(z). Along with a breakdown by commercial/non-commercial traders, the CFTC also reports 'disaggregated' categories of producer/merchant, swap dealers, managed money and other reportables. When a contract does not meet the CFTC's minimum reporting requirement there will be no data reported and that date will be left blank.

SOURCE: http://www.cftc.gov/MarketReports/CommitmentsofTraders/index.htm

UPDATE INFO: Data is provided weekly on Friday at 3:30pm EST for the week ending on Tuesday. Data is only reported to the CFTC for markets in which 20 or more traders hold positions equal to or above the reporting levels established by the CFTC. If the minumum requirements are not met, no data will be posted for that week.

CALCULATIONS: Data is taken directly from the source without any manipulation, except for the 'Net' tickers, which are calculated by taking the long contracts minus the short contracts.

SOURCE METHODOLOGY: http://www.cftc.gov/MarketReports/CommitmentsofTraders/ExplanatoryNotes/index.htm

Please note that "The Disaggregated COT report data for physical commodity markets can be re-aggregated to get back to the two categories of the COT report. The TFF report, however, is not a disaggregation of the COT data for the financial futures markets. The traders classified into one of the four categories in the TFF report may be drawn from either the 'commercial' or 'noncommercial' categories of traders in the legacy COT reports."

HISTORICAL CONSIDERATIONS: The Coffee Sugar and Cocoa Exchange and New York Cotton Exchange were part of the New York Board of Trade which merged with Intercontinental Exchange (ICE) in January 2007. If a holiday falls on a Tuesday, report dates for the Monday before or Wednesday after are modified to reflect the Tuesday date, as if there were no holiday.

DATA DISCOVERY: For more details see COT<GO> <HELP> on a Bloomberg terminal.

RELATED SECTORS/FUNCTIONS: hit FDM <GO> on a Bloomberg terminal.

TFF Traders in Financial Futures Report:
Explanatory notes: http://www.cftc.gov/idc/groups/public/@commitmentsoftraders/documents/file/tfmexplanatorynotes.pdf
Future's only: hit ALLX TFF1<GO> ALLX TFF2<GO> ALLX TFF3<GO> on a Bloomberg terminal.
Future's & options combined: hit ALLX TFC1<GO> ALLX TFC2<GO> ALLX TFC3<GO> on a Bloomberg terminal.

---CFTC ticker guide---

Ticker ends with:

Legacy:

'COL': Commercial Long Contracts.

'COS': Commercial Short Contracts.

'CON': Commercial Net Contracts.

'NCL': Non-Commercial Long Contracts.

'NCP': Non-Commercial Spreading Contracts.

'NCS': Non-Commercial Short Contracts.

'NCN': Non-Commercial Net Contracts.

'NRL': Non-Reportable Long Contracts.

'NRS': Non-Reportable Short Contracts.

'NRN': Non-Reportable Net Contracts.

'OIN': Total Open Interest.

'TCL': Commercial Long Traders.

'TCS': Commercial Short Traders.

'TLL': Total Long Contracts.

'TLS': Total Short Contracts.

'TLN': Total Net Contracts.

'TNL': Non-Commercial Long Traders.

'TNP': Non-Commercial Spreading Traders.

'TNS': Non-Commercial Short Traders.

'TTL': Total Long Traders.

'TTO': Total Traders.

'TTS': Total Short Traders.

Disaggregated:

'PML': Producer Merchant Long Contracts.

'PMS': Producer Merchant Short Contracts.

'PMN': Producer Merchant Net Contracts.

'PTL': Producer Merchant Total Long Traders.

'PTS': Producer Merchant Total Short Traders.

'SWL': Swap Dealers Long Contracts.

'SWD': Swap Dealers Spreading Contracts.

'SWS': Swap Dealers Short Contracts.

'SWN': Swap Dealers Net Contracts.

'STL': Swap Dealers Total Long Traders.

'STS': Swap Dealers Total Short Traders.

'STD': Swap Dealers Total Spread Traders.

'MML': Managed Money Long Contracts.

'MMD': Managed Money Spreading Contracts.

'MMS': Managed Money Short Contracts.

'MMN': Managed Money Net Contracts.

'MTL': Managed Money Total Long Traders.

'MTS': Managed Money Total Short Traders.

'MTD': Managed Money Total Spread Traders.

'ORL': Other Reportables Long Contracts.

'ORD': Other Reportables Spreading Contracts.

'ORS': Other Reportables Short Contracts.

'ORN': Other Reportables Net Contracts.

'OTL': Other Reportables Total Long Traders.

'OTS': Other Reportables Total Short Traders.

'OTD': Other Reportables Total Spread Trader.

Traders in Financial Futures (TFF):

'DIL': Dealer Intermediary Long Contracts.

'DIS': Dealer Intermediary Short Contracts.

'DIN': Dealer Intermediary Net Total.

'DID': Dealer Intermediary Spread Contracts.

'DTL': Dealer Intermediary Total Long Traders.

'DTS': Dealer Intermediary Total Short Traders.

'DTD': Dealer Intermediary Total Spread Traders.

'AIL': Asst Mgr Institutional Long Contracts.

'AIS': Asst Mgr Institutional Short Contracts.

'AIN': Asst Mgr Institutional Net Totals.

'AID': Asst Mgr Institutional Spread Contracts.

'ATL': Asst Mgr Institutional Total Long Traders.

'ATS': Asst Mgr Institutional Total Short Traders.

'ATD': Asst Mgr Institutional Total Spread Traders.

'LFL': Leveraged Funds Long Contracts.

'LFS': Leveraged Funds Short Contracts.

'LFN': Leveraged Funds Net Totals.

'LFD': Leveraged Funds Spread Contracts.

'LTL': Leveraged Funds Total Long Traders.

'LTS': Leveraged Funds Total Short Traders.

'LTD': Leveraged Funds Total Spread Traders.

'ORL': Other Reportables Long Contracts.

'ORS': Other Reportables Short Contracts.

'ORN': Other Reportables Net Totals.

'ORD': Other Reportables Spread Contracts.

'OTL': Other Reportables Total Long Traders.

'OTS': Other Reportables Total Short Traders.

'OTD': Other Reportables Total Spread Traders.

Supplemental/Commodity Index Traders (CIT):

'CLN': Non Commerical Long Non CIT.

'CSN': Non Commerical Short Non CIT.

'NPN': Non Commerical Spreading Non CIT.

'NNT': Non Commerical Non CIT (Net).

'OLN': Commerical Long Non CIT.

'OSN': Commerical Short Non CIT.

'CNT': Commerical Non CIT (Net).

'CIL': Index Trader Long.

'CIS': Index Trader Short.

'CIN': Index Trader (Net).

'NLN': # Traders Non Comm Long Non CIT.

'NSN': # Traders Non Comm Short Non CIT.

'STN': # Traders Non Comm Spread Non CIT.

'LTN': # Traders Comm Long Non CIT.

'TSN': # Traders Comm Short Non CIT.

'ILT': # Traders Index Trader Long CIT.

'IST': # Traders Index Trader Short CIT.

'TLC': # Traders Index Trader Long Non CIT.

'SNT': # Traders Index Trader Short Non CIT.

'LTS': Leveraged Funds Total Short Traders.

'LTD': Leveraged Funds Total Spread Traders.

'ORL': Other Reportables Long Contracts.

'ORS': Other Reportables Short Contracts.

'ORN': Other Reportables Net Totals.

'ORD': Other Reportables Spread Contracts.

'OTL': Other Reportables Total Long Traders.

'OTS': Other Reportables Total Short Traders.

'OTD': Other Reportables Total Spread Traders.

This data represents the number of non-commerical long traders who must report their positions for the stated futures and/or option contracts. A trader must report his or her postion if at the daily market close, their position is at or above the CFTC's reporting level in any futures month or option expiration. A trader is determined to be commercial or non-commerical using the following rationale: all trader's reported futures positions in a commodity are classifed as commercial if the trader uses futures contracts in that particular commodity for hedging as defined in the CFTC's regulations (1.3(z)).

See also