Bloomberg tickers for various popular futures series

The dataset includes Bloomberg active contract tickers for several popular futures series including commodity, currency, financial as well as index futures with underlyings from various asset classes.

tickers_futures

Format

A data.table. Columns include:

ticker: active contract ticker for the futures series.

name: corresponding futures series's underlying name.

asset class: asset class for the underlying futures series; 'commodity', 'currency', 'financial' and 'index'.

sector: underlying instrument sector; asset class specific.

commodity: 'agriculturals', 'energy', 'metals'.

currency: domestic currency.

financial: instrument type.

index: index asset class.

subsector: underlying instrument subsector; asset class specific.

commodity:

agriculturals: 'grains', 'livestock', 'softs'.

energy: 'gas', 'petroleum'.

metals: 'base', 'precious'.

currency: foreign currency.

financial: for government debt underlyings, corresponding country code (ISO 3166-1 alpha-2).

index. sector specific:

commodity indexes (commodity): commodity sector covered by the index.

equity indexes (equity): country/region covered by the index.

currency: currency for observed price variables.

MIC: ISO 10383 Codes for exchanges and market identification (MIC). Identifies the exchange where the corresponding futures series trades.

term structure length: length of the term structure for the futures series (number of contracts).

contract size: size of the corresponding contract with units specified in trading unit below.

point value: change in contract value for one point change in price.

tick size: minimum price fluctuation.

tick value: change in price resulting for a minimum price fluctuation described above.

FIGI: Financial Instrument Global Identifier. Twelve character, alphanumeric identifier. The first 2 characters are upper-case consonants (including "Y"), the third character is the upper-case "G", characters 4 -11 are any upper-case consonant (including "Y") or integer between 0 and 9, and the last character is a check-digit. An identifier is assigned to instruments of all asset classes, is unique to an individual instrument and once issued will not change for an instrument. For equity instruments an identifier is issued per instrument per trading venue.

description: descriptive elements for the corresponding futures series.

Source

Bloomberg

See also